Changing publisher components directory and dto classes name
