+++ OrderService/Infrastructure/ComplementaryServices/RabbitMQ/Publishers/Contracts/IPublisherQueue.cs
@@ -1,0 +1,10 @@
+﻿using Infrastructure.ComplementaryServices.RabbitMQ.Publishers.Settings;
+
+namespace Infrastructure.ComplementaryServices.RabbitMQ.Publishers.Contracts
+{
+    public interface IPublisherQueue
+    {
+        bool IsMatch(AvailableQueue availableQueue);
+        bool Publish(string jsonContent);
+    }
+}
