+++ OrderService/Infrastructure/ComplementaryServices/RabbitMQ/Publishers/Settings/AvailableQueue.cs
@@ -1,0 +1,7 @@
+﻿namespace Infrastructure.ComplementaryServices.RabbitMQ.Publishers.Settings
+{
+    public enum AvailableQueue
+    {
+        CreditCard = 1
+    }
+}
