--- OrderService/Infrastructure/External/RabbitMQ/Publishers/Contracts/IPublisherQueue.cs
@@ -1,10 +1,0 @@
-﻿using Infrastructure.External.RabbitMQ.Publishers.Settings;
-
-namespace Infrastructure.External.RabbitMQ.Publishers.Contracts
-{
-    public interface IPublisherQueue
-    {
-        bool IsMatch(AvailableQueue availableQueue);
-        bool Publish(string jsonContent);
-    }
-}
