--- OrderService/Infrastructure/External/RabbitMQ/Publishers/Settings/AvailableQueue.cs
@@ -1,7 +1,0 @@
-﻿namespace Infrastructure.External.RabbitMQ.Publishers.Settings
-{
-    public enum AvailableQueue
-    {
-        CreditCard = 1
-    }
-}
